+++ CHANGES.rst
@@ -10,9 +10,8 @@ Unreleased
 - Refuse usage of pytest versions 4.3, 4.4
 - Added tags support for Examples sections
 - Added joining by parameters between examples sections on different levels (and with fixtures)
-- Parsed step parameters are not fixtures anymore. Hook **pytest_bdd_get_parameter_context_value** could be used to get compiled parameter for current step
-- Step could override multiple fixtures using **target_fixtures** parameter
-
+- Step could override multiple fixtures using ``target_fixtures`` parameter
+- Default step parameters injection as fixtures behavior could be changed by ``params_fixtures_mapping`` step parameter
 
 5.0.0
 -----